search refactor
