+++ packages/client/src/app/components/modals/trading/orderbook/controls/index.ts
@@ -1,0 +1,1 @@
+export { Controls } from './Controls';
